método para ação registrada e correção dto

+++ src/main/java/rei_da_quadra_be/dto/PartidaResponseDTO.java
@@ -10,9 +10,9 @@ public class PartidaResponseDTO {
   private String status;
   private String timeANome;
   private Integer timeAPlacar;
-  private Integer timeAId;
+  private Long timeAId;
   private String timeBNome;
   private Integer timeBPlacar;
-  private Integer timeBId;
+  private Long timeBId;
   private Long eventoId;
 }
